Add '--include-hidden-tasks' option to enhance '-T' and '-D'
Since we can invoke tasks that do not have descriptions, we ought to be
able to display and describe them, too.

The new '--include-hidden-tasks' command-line option has no effect by
itself. But when it is specified together with '-T' ('--tasks') or '-D'
('--describe'), it causes Rake to display tasks that lack descriptions
as well as those that have descriptions.